implement update shopify dataset description
